pyproject.toml
[project]
name = "q-bot"
version = "0.1.0"
description = "q-bot"
readme = "README.md"
requires-python = ">=3.8, <4.0"

[tool.nonebot]
adapters = [
    { name = "OneBot V11", module_name = "nonebot.adapters.onebot.v11" }
]
# ,"nonebot_plugin_ikun_evolution"
# "nonebot_plugin_gocqhttp",
# "nonebot_plugin_status", 
plugins = ["nonebot_plugin_apscheduler", "nonebot_plugin_guild_patch", "nonebot_plugin_tortoise_orm"]
plugin_dirs = ["src/plugins"]
# builtin_plugins = ["echo"]

[tool.poetry]
name = "q-bot"
version = "0.1.0"
description = ""
authors = ["mehaotian <490272692@qq.com>"]
readme = "README.md"
packages = [{include = "q_bot"}]

[tool.poetry.dependencies]
python = "^3.10"
nonebot2 = {extras = ["fastapi"], version = "^2.0.0"}
nonebot-adapter-onebot = "^2.2.3"
nonebot-plugin-gocqhttp = "^0.6.10"
nonebot-plugin-status = "^0.7.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
